Rename testtrace to tracetest
