+++ api/trace/tracetest/trace_context_propagator_benchmark_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testtrace_test
+package tracetest_test
 
 import (
 	"context"
